feat: modify AbstractConfig
