+++ arex-storage-web-api/src/main/java/com/arextest/storage/service/config/AbstractConfig.java
@@ -1,5 +1,7 @@
 package com.arextest.storage.service.config;
 
+import org.apache.commons.lang3.StringUtils;
+
 /**
  * @author niyan
  * @date 2024/6/14
@@ -9,7 +11,20 @@ public abstract class AbstractConfig {
 
     protected abstract String getConfigAsString(String key);
 
-    public abstract String getConfigAsString(String key, String defaultValue);
+    public String getConfigAsString(String key, String defaultValue) {
+        return StringUtils.defaultString(getConfigAsString(key), defaultValue);
+    }
 
-    public abstract int getConfigAsInt(String key, int defaultValue);
+    public int getConfigAsInt(String key, int defaultValue) {
+        String value = getConfigAsString(key);
+        if (!StringUtils.isNumeric(value)) {
+            return defaultValue;
+        }
+
+        try {
+            return Integer.parseInt(value);
+        } catch (NumberFormatException e) {
+            return defaultValue;
+        }
+    }
 }
